fix: removed circular deps
