+++ src/module/item/process.ts
@@ -1,5 +1,5 @@
 import { FifthItem, ItemRarity } from '../fifthedition.actor.template';
-import { parsedToWeapon } from './weapon';
+import { ItemType, parsedToWeapon, parseType } from './weapon';
 
 interface ItemData {
   description: {
@@ -7,8 +7,6 @@ interface ItemData {
   };
   rarity: ItemRarity;
 }
-
-type ItemType = 'weapon' | 'equipment' | 'consumable' | 'tool' | 'loot' | 'class' | 'spell' | 'feat' | 'backpack';
 
 export interface Item {
   name: string;
@@ -20,18 +18,6 @@ function parseName(input: string): string {
   return input.split('\n')[0].trim();
 }
 
-export function parseType(input: string): ItemType {
-  if (/weapon/i.test(input)) return 'weapon';
-  if (/armor/i.test(input)) return 'equipment';
-  if (/unil the next dawn/i.test(input)) return 'consumable';
-  if (/beginning at/i.test(input)) return 'feat';
-  if (/starting at/i.test(input)) return 'feat';
-  if (/melee weapon attack/i.test(input)) return 'weapon';
-  if (/ranged weapon attack/i.test(input)) return 'weapon';
-  if (/melee or ranged weapon attack/i.test(input)) return 'weapon';
-  return 'consumable';
-}
-
 function parseRarity(input: string): ItemRarity {
   if (/uncommon/i.test(input)) return 'uncommon';
   if (/common/i.test(input)) return 'common';
@@ -40,10 +26,14 @@ function parseRarity(input: string): ItemRarity {
   return 'common';
 }
 
+function scrubDescription(description: string, extraItems: string[]): string {
+  return extraItems.reduce((acc, item) => acc.replace(item, '').trim(), description).replace('\n', '');
+}
+
 export function processItem(input: string): FifthItem {
   const type = parseType(input);
   const name = parseName(input);
-  const description = input.replace(name, '').trim();
+  const description = scrubDescription(input, [name, type]);
   switch (type) {
     case 'weapon':
     case 'feat':
