+++ src/module/item/weapon.ts
@@ -1,6 +1,28 @@
 import { Abilities, Feature, parseFormula } from '../actors.process';
 import { ActionType, FifthFeatureCost, FifthItem, FifthItemType, FifthStat } from '../fifthedition.actor.template';
-import { parseType } from './process';
+
+export type ItemType =
+  | 'weapon'
+  | 'equipment'
+  | 'consumable'
+  | 'tool'
+  | 'loot'
+  | 'class'
+  | 'spell'
+  | 'feat'
+  | 'backpack';
+
+export function parseType(input: string): ItemType {
+  if (/weapon/i.test(input)) return 'weapon';
+  if (/armor/i.test(input)) return 'equipment';
+  if (/unil the next dawn/i.test(input)) return 'consumable';
+  if (/beginning at/i.test(input)) return 'feat';
+  if (/starting at/i.test(input)) return 'feat';
+  if (/melee weapon attack/i.test(input)) return 'weapon';
+  if (/ranged weapon attack/i.test(input)) return 'weapon';
+  if (/melee or ranged weapon attack/i.test(input)) return 'weapon';
+  return 'consumable';
+}
 
 function getDamageType(from: string): string | undefined {
   if (from.includes('piercing')) return 'piercing';
